refactor: code simplification

+++ record-stream.sh
@@ -3,8 +3,7 @@ set -eo pipefail
 
 is_streaming() {
     [ "$DEBUG" = true ] && [ -e "is_streaming" ] && return 0
-    local response=$(curl -s https://r-a-d.io/api | jq -r '.main.isafkstream')
-    [[ "$response" != "true" ]]
+    curl -s https://r-a-d.io/api | jq -e '.main.isafkstream == false' >/dev/null
 }
 
 is_recording() {
@@ -12,40 +11,35 @@ is_recording() {
     [[ -n "$pid" ]] && [[ -n $(ps -p $pid -o pid=) ]]
 }
 
+start_recording() {
+    recording="r-a-dio_$(date +"%Y-%m-%dT%H-%M-%S-%3N").mp3"
+    log="${recording}.log"
+    wget https://stream.r-a-d.io/main.mp3 -O "$recording" >/dev/null 2>&1 &
+    pid=$!
+}
+
+stop_recording() {
+    [ -n "$pid" ] && kill "$pid" || true
+}
+
 append_lp_log() {
-    local lp_file="$1"
-    curl -s https://r-a-d.io/api | jq -r '.main.lp[] | "\(.timestamp): \(.meta)"' >>"$lp_file"
-    sort -u -o "$lp_file" "$lp_file"
+    curl -s https://r-a-d.io/api | jq -r '.main.lp[] | "\(.timestamp): \(.meta)"' >>"$log"
+    sort -u -o "$log" "$log"
 }
 
-cleanup() {
-    if [[ -n $pid ]]; then
-        echo "Killing process $pid..."
-        kill "$pid" || true
-    fi
-}
-
-trap cleanup EXIT
+trap stop_recording EXIT
 
 while true; do
-    if is_streaming && ! is_recording; then
-        echo "Recording started."
-
-        recording="r-a-dio_$(date +"%Y-%m-%dT%H-%M-%S-%3N").mp3"
-        log="${recording}.log"
-        touch "$log"
-
-        wget https://stream.r-a-d.io/main.mp3 -O "$recording" >/dev/null 2>&1 &
-        pid=$!
-
-        append_lp_log "$log"
-    elif is_streaming && is_recording; then
-        append_lp_log "$log"
-    elif ! is_streaming && is_recording; then
-        echo "Recording stopped."
-        append_lp_log "$log"
-        kill "$pid" || true
-        pid=""
+    if ! is_recording && is_streaming; then
+        start_recording
+        echo "Recording started: $recording"
+        append_lp_log
+    elif is_recording && ! is_streaming; then
+        stop_recording
+        echo "Recording stopped: $recording"
+        append_lp_log
+    elif is_recording && is_streaming; then
+        append_lp_log
     fi
     sleep 60
 done
